admin can delete messages and remove users from channels, user disconnecting also emits to channels, so the user is properly removed everywhere, added channel statistics and optional password to channels

+++ src/socket-handlers/channel.handler.ts
@@ -1,7 +1,8 @@
-import { Socket, Server } from 'socket.io'
+import { Socket, RemoteSocket, Server } from 'socket.io'
 import Channel, { ChannelDocument } from '../models/channel.model'
-import { UserDocument, UserRoles } from '../models/user.model'
+import { UserDocument } from '../models/user.model'
 import Message from '../models/message.model'
+import { isAdmin } from '../utils/user-utils'
 
 import socketUserMap from '../socket/socket-map'
 
@@ -10,10 +11,10 @@ export const channelHandler = (io: Server, socket: Socket) => {
   socket.on('one:channel:delete', async (channel: ChannelDocument) => {
     const user: UserDocument = socket.data.user
 
-    // if (user.role !== UserRoles.Admin) {
-    //   socket.emit('unauthorized', { message: 'User doesnt have proper permissions to delete a channel' })
-    //   return
-    // }
+    if (!isAdmin(user)) {
+      socket.emit('unauthorized', { message: 'User doesnt have proper permissions to delete a channel' })
+      return
+    }
 
     const { alias } = channel
     await Message.find({ channel: channel._id }).deleteMany()
@@ -39,7 +40,7 @@ export const channelHandler = (io: Server, socket: Socket) => {
     console.log(channelUsers, 'CHANNEL USERS')
 
     socket.emit('one:channel:joined', { messages: messages.reverse(), channel, channelUsers })
-    io.to(channel.alias).emit('all:channel:joined', {
+    io.to(channel.alias).emit('all:channel:user:joined', {
       alias,
       socketId: socket.id
     })
@@ -47,20 +48,36 @@ export const channelHandler = (io: Server, socket: Socket) => {
 
   socket.on('one:channel:leave', (channel: ChannelDocument) => {
     socket.leave(channel.alias)
-    io.to(channel.alias).emit('all:channel:left', {
+    io.to(channel.alias).emit('all:channel:user:left', {
       alias: channel.alias,
       socketId: socket.id
     })
   })
 
-  // socket.on('all:channel:left', (channel: ChannelDocument) => {
-  //   console.log('LEAVE channel', channel)
-  //   socket.leave(channel.alias)
-  //   io.to(channel.alias).emit('all:channel:left', {
-  //     alias: channel.alias,
-  //     socketId: socket.id
-  //   })
-  // })
+  socket.on('one:channel:user:remove', async ({ socketId, alias }) => {
+    const user: UserDocument = socket.data.user
+    if (!isAdmin(user)) {
+      socket.emit('unauthorized', { message: 'User doesnt have proper permissions to remove user from channel' })
+      return
+    }
+
+    const userSockets = await io.in(alias).fetchSockets()
+    const userSocket = userSockets.find((socket) => socket.id.toString() === socketId)
+
+    if (!userSocket) {
+      throw new Error('Invalid socket id received')
+    }
+
+    userSocket.leave(alias)
+    userSocket.emit('one:channel:removed', { alias })
+    io.to(alias).emit('all:channel:user:left', { alias, socketId })
+  })
+
+  socket.on('one:channel:message:delete', async ({ alias, message }) => {
+    const { _id } = message
+    await Message.deleteOne({ _id })
+    io.to(alias).emit('all:channel:message:deleted', { alias, message })
+  })
 
   socket.on('one:channel:message:sent', async ({ alias, channelId, userId, message }) => {
     const newMessage = new Message({
